+++ tests/testthat/test_issue81.R
@@ -2,14 +2,17 @@
 ## excel_repeat.ods is created with MS Office 365 online
 
 test_that("issue 81, correctness", {
-    res <- read_ods("../testdata/excel_repeat.ods", col_names = FALSE)
+    res <- read_ods("../testdata/excel_repeat.ods", col_names = FALSE, as_tibble = FALSE)
     expect_identical(res[,1], c(rep("A", 12), rep("C", 11)))
     expect_identical(res[,2], c(rep("B", 12), rep("D", 11)))
+    res <- read_ods("../testdata/excel_repeat.ods", col_names = FALSE)
+    expect_identical(res[[1]], c(rep("A", 12), rep("C", 11)))
+    expect_identical(res[[2]], c(rep("B", 12), rep("D", 11)))
 })
 
 test_that("issue 81 real test", {
     file <- "../testdata/issue81.ods"
-    res <- read_ods(file, sheet = 2, skip = 4)
+    res <- read_ods(file, sheet = 2, skip = 4, as_tibble = FALSE)
     expect_equal(sum(is.na(res[,1])), 0)
 })