+++ tests/testthat/test_write_ods_append_update.R
@@ -39,36 +39,36 @@ test_that("Write Excel sheets", {
     expect_error(write_ods(starwars10, tmp, "SWRC", row_names=TRUE, col_names = TRUE, update = TRUE), NA)
     expect_error(write_ods(starwars10, tmp, "whatevernotexists", row_names=TRUE, col_names = TRUE, update = TRUE))
 
-    df <- read_ods(tmp, "SW", row_names = FALSE, col_names = FALSE, strings_as_factors = TRUE)
+    df <- read_ods(tmp, "SW", row_names = FALSE, col_names = FALSE, strings_as_factors = TRUE, as_tibble = FALSE)
     expect_true(all.equal({
         cars <- starwars10
         rownames(cars) <- NULL
-        colnames(cars) <- cols_to_letters(ncol(cars))
+        colnames(cars) <- vctrs::vec_as_names(rep("", 9), repair = "unique")
         cars
     }, df))
 
-    df <- read_ods(tmp, "SWR", row_names = TRUE, col_names = FALSE, strings_as_factors = TRUE)
+    df <- read_ods(tmp, "SWR", row_names = TRUE, col_names = FALSE, strings_as_factors = TRUE, as_tibble = FALSE)
     expect_true(all.equal({
         cars <- starwars10
-        colnames(cars) <- cols_to_letters(ncol(cars))
+        colnames(cars) <- vctrs::vec_as_names(rep("", 9), repair = "unique")
         cars}, df))
 
-    df <- read_ods(tmp, "SWC", row_names = FALSE, col_names = TRUE, strings_as_factors = TRUE)
+    df <- read_ods(tmp, "SWC", row_names = FALSE, col_names = TRUE, strings_as_factors = TRUE, as_tibble = FALSE)
     expect_true(all.equal({
         cars <- starwars10
         rownames(cars) <- NULL
         cars
     }, df))
 
-    df <- read_ods(tmp, "SWRC", row_names = TRUE, col_names = TRUE, strings_as_factors = TRUE)
+    df <- read_ods(tmp, "SWRC", row_names = TRUE, col_names = TRUE, strings_as_factors = TRUE, as_tibble = FALSE)
     expect_true(all.equal(starwars10, df))
 
-    df <- read_ods(tmp, "SW1", row_names = TRUE, col_names = TRUE, strings_as_factors = TRUE)
+    df <- read_ods(tmp, "SW1", row_names = TRUE, col_names = TRUE, strings_as_factors = TRUE, as_tibble = FALSE)
 
     expect_false(isTRUE(all.equal(starwars10[1, seq_len(ncol(starwars10))], df))) # factor mismatch
     expect_true(all((df == starwars10[1, seq_len(ncol(starwars10))])[1,]))
 
-    df <- read_ods(tmp, "SW10", row_names = TRUE, col_names = TRUE, strings_as_factors = TRUE)
+    df <- read_ods(tmp, "SW10", row_names = TRUE, col_names = TRUE, strings_as_factors = TRUE, as_tibble = FALSE)
     expect_true(all.equal(starwars10[seq_len(nrow(starwars10)), 1, drop=FALSE], df))
 })
 
@@ -77,4 +77,11 @@ test_that("issue 107", {
     expect_error(write_ods(legend, tmp, sheet = "Legend"), NA)
     expect_error(write_ods(legend, tmp, sheet = "Legend", update = TRUE), NA)
     expect_error(write_ods(legend, tmp, sheet = "Legend2", append = TRUE), NA)
+})
+
+test_that("reading and writing and reading gets the same result as the start", {
+    a <- read_ods("../testdata/starwars.ods")
+    b <- write_ods(a, tmp)
+    c <- read_ods(tmp)
+    expect_equal(a,c)
 })